change input output path for PR1954

+++ Preprocessing/RemoveAdPage.sh
@@ -8,11 +8,16 @@ declare -a AdPage=("pr1956_f0063_2_1.json"
                         "pr1956_f0128_2_1.json"
                         "pr1956_f0128_3_0.json" )
 
-Path=${Path:-'../../personnel-records/1956/seg/page_rect/'}
+Path=${Path:-'../../results/personnel-records/1954/seg/page_rect/'}
 
-echo "Removing Advertisement Page"
+read -p "Do you want to remove advertisement page in $OutputPath? (y/n) " -n 1 -r
+echo -e "\n"
+if [[ $REPLY =~ ^[Yy]$ ]]
+then
+    echo "Removing Advertisement Page"
 
-for json in "${AdPage[@]}"; do
-  echo "Removing ${Path}$json"
-  rm ${Path}$json
-done
+    for json in "${AdPage[@]}"; do
+      echo "Removing ${Path}$json"
+      rm ${Path}$json
+    done
+if
